separate folder for schemas
