+++ routes/api/contacts.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
 
-const validation = require("../../helpers/validation");
+const validation = require("../../schemas/validation");
 const controllers = require("../../controllers/controllers");
 
 router.get('/', controllers.getContacts)
